chore: Refactored GlobalFaceting into folder with base, types, and utils
