+++ packages/table-core/src/core/table.ts
@@ -30,7 +30,7 @@ import { ColumnOrdering } from '../features/column-ordering/ColumnOrdering'
 import { ColumnPinning } from '../features/column-pinning/ColumnPinning'
 import { ColumnSizing } from '../features/column-sizing/ColumnSizing'
 import { ColumnVisibility } from '../features/column-visibility/ColumnVisibility'
-import { GlobalFaceting } from '../features/GlobalFaceting'
+import { GlobalFaceting } from '../features/global-faceting/GlobalFaceting'
 import { GlobalFiltering } from '../features/GlobalFiltering'
 import { RowExpanding } from '../features/RowExpanding'
 import { RowPagination } from '../features/RowPagination'
