+++ packages/table-core/src/features/global-faceting/GlobalFaceting.types.ts
@@ -1,5 +1,4 @@
-import { RowModel } from '..'
-import { Table, RowData, TableFeature } from '../types'
+import { RowData, RowModel } from '../../types'
 
 export interface GlobalFacetingInstance<TData extends RowData> {
   _getGlobalFacetedMinMaxValues?: () => undefined | [number, number]
@@ -23,44 +22,4 @@ export interface GlobalFacetingInstance<TData extends RowData> {
    * @link [Guide](https://tanstack.com/table/v8/docs/guide/global-faceting)
    */
   getGlobalFacetedUniqueValues: () => Map<any, number>
-}
-
-//
-
-export const GlobalFaceting: TableFeature = {
-  _createTable: <TData extends RowData>(table: Table<TData>): void => {
-    table._getGlobalFacetedRowModel =
-      table.options.getFacetedRowModel &&
-      table.options.getFacetedRowModel(table, '__global__')
-
-    table.getGlobalFacetedRowModel = () => {
-      if (table.options.manualFiltering || !table._getGlobalFacetedRowModel) {
-        return table.getPreFilteredRowModel()
-      }
-
-      return table._getGlobalFacetedRowModel()
-    }
-
-    table._getGlobalFacetedUniqueValues =
-      table.options.getFacetedUniqueValues &&
-      table.options.getFacetedUniqueValues(table, '__global__')
-    table.getGlobalFacetedUniqueValues = () => {
-      if (!table._getGlobalFacetedUniqueValues) {
-        return new Map()
-      }
-
-      return table._getGlobalFacetedUniqueValues()
-    }
-
-    table._getGlobalFacetedMinMaxValues =
-      table.options.getFacetedMinMaxValues &&
-      table.options.getFacetedMinMaxValues(table, '__global__')
-    table.getGlobalFacetedMinMaxValues = () => {
-      if (!table._getGlobalFacetedMinMaxValues) {
-        return
-      }
-
-      return table._getGlobalFacetedMinMaxValues()
-    }
-  },
 }
